[Refactor] dto 책임 분리
