+++ src/careworker/dto/get-careworkers.dto.ts
@@ -1,0 +1,41 @@
+import { ApiProperty } from '@nestjs/swagger';
+
+export class GetCareworkersResponseDto {
+  @ApiProperty({
+    description: '요청 성공 여부',
+    example: true,
+  })
+  success: boolean;
+
+  @ApiProperty({
+    description: '요양보호사 목록',
+    type: 'array',
+    items: {
+      type: 'object',
+      properties: {
+        id: { type: 'number', example: 1 },
+        email: { type: 'string', example: 'careworker@example.com' },
+        name: { type: 'string', example: '김요양' },
+        role: { type: 'string', example: 'careWorker' },
+        createdAt: {
+          type: 'string',
+          format: 'date-time',
+          example: '2024-01-01T00:00:00.000Z',
+        },
+        updatedAt: {
+          type: 'string',
+          format: 'date-time',
+          example: '2024-01-01T00:00:00.000Z',
+        },
+      },
+    },
+  })
+  data: {
+    id: number;
+    email: string;
+    name: string;
+    role: string;
+    createdAt: Date;
+    updatedAt: Date;
+  }[];
+}
